+++ packages/isaacscript-common/src/objects/roomShapeCorners.ts
@@ -2,10 +2,12 @@ import { RoomShape } from "isaac-typescript-definitions";
 import { CornerType } from "../enums/CornerType";
 import { Corner } from "../interfaces/Corner";
 
-const TWO_BY_TWO_CENTER_X = 580;
-const TWO_BY_TWO_CENTER_Y = 420;
-
-/** The locations of the corners for each room shape. */
+/**
+ * The locations of the corners for each room shape.
+ *
+ * Note that these corner locations are not accurate for the Mother Boss Room and the Home closet
+ * rooms. (Those rooms have custom shapes.)
+ */
 export const ROOM_SHAPE_CORNERS: {
   readonly [key in RoomShape]: readonly Corner[];
 } = {
@@ -13,18 +15,22 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.SHAPE_1x1]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 0,
       position: Vector(60, 140),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 14,
       position: Vector(580, 140),
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 120,
       position: Vector(60, 420),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 134,
       position: Vector(580, 420),
     },
   ],
@@ -33,18 +39,22 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.IH]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 30,
       position: Vector(60, 220),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 44,
       position: Vector(580, 220),
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 90,
       position: Vector(60, 340),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 104,
       position: Vector(580, 340),
     },
   ],
@@ -53,18 +63,22 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.IV]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 4,
       position: Vector(220, 140),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 10,
       position: Vector(420, 140),
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 124,
       position: Vector(220, 420),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 130,
       position: Vector(420, 420),
     },
   ],
@@ -73,18 +87,22 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.SHAPE_1x2]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 0,
       position: Vector(60, 140),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 14,
       position: Vector(580, 140),
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 225,
       position: Vector(60, 700),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 239,
       position: Vector(580, 700),
     },
   ],
@@ -93,18 +111,22 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.IIV]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 4,
       position: Vector(220, 140),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 10,
       position: Vector(420, 140),
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 229,
       position: Vector(220, 700),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 235,
       position: Vector(420, 700),
     },
   ],
@@ -113,18 +135,22 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.SHAPE_2x1]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 0,
       position: Vector(60, 140),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 27,
       position: Vector(1100, 140),
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 224,
       position: Vector(60, 420),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 251,
       position: Vector(1100, 420),
     },
   ],
@@ -133,18 +159,22 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.IIH]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 56,
       position: Vector(60, 220),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 83,
       position: Vector(1100, 220),
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 168,
       position: Vector(60, 340),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 195,
       position: Vector(1100, 340),
     },
   ],
@@ -153,18 +183,22 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.SHAPE_2x2]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 0,
       position: Vector(60, 140),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 27,
       position: Vector(1100, 140),
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 420,
       position: Vector(60, 700),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 447,
       position: Vector(1100, 700),
     },
   ],
@@ -173,26 +207,32 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.LTL]: [
     {
       type: CornerType.TOP_LEFT,
-      position: Vector(TWO_BY_TWO_CENTER_X, 140), // Center
+      gridIndex: 13,
+      position: Vector(580, 140), // Center
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 27,
       position: Vector(1100, 140),
     },
     {
       type: CornerType.TOP_LEFT,
-      position: Vector(60, TWO_BY_TWO_CENTER_Y), // Left
+      gridIndex: 196,
+      position: Vector(60, 420), // Left
     },
     {
       type: CornerType.BOTTOM_RIGHT,
-      position: Vector(TWO_BY_TWO_CENTER_X, TWO_BY_TWO_CENTER_Y), // Center
+      gridIndex: 209,
+      position: Vector(580, 420), // Center
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 420,
       position: Vector(60, 700),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 447,
       position: Vector(1100, 700),
     },
   ],
@@ -201,26 +241,32 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.LTR]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 0,
       position: Vector(60, 140),
     },
     {
       type: CornerType.TOP_RIGHT,
-      position: Vector(TWO_BY_TWO_CENTER_X, 140), // Center
+      gridIndex: 14,
+      position: Vector(580, 140), // Center
     },
     {
       type: CornerType.BOTTOM_LEFT,
-      position: Vector(TWO_BY_TWO_CENTER_X, TWO_BY_TWO_CENTER_Y), // Center
+      gridIndex: 210,
+      position: Vector(580, 420), // Center
     },
     {
       type: CornerType.TOP_RIGHT,
-      position: Vector(1100, TWO_BY_TWO_CENTER_Y), // Right
+      gridIndex: 223,
+      position: Vector(1100, 420), // Right
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 420,
       position: Vector(60, 700),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 447,
       position: Vector(1100, 700),
     },
   ],
@@ -229,26 +275,32 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.LBL]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 0,
       position: Vector(60, 140),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 27,
       position: Vector(1100, 140),
     },
     {
       type: CornerType.BOTTOM_LEFT,
-      position: Vector(TWO_BY_TWO_CENTER_X, 140), // Left
+      gridIndex: 224,
+      position: Vector(580, 140), // Left
     },
     {
       type: CornerType.TOP_RIGHT,
-      position: Vector(TWO_BY_TWO_CENTER_X, TWO_BY_TWO_CENTER_Y), // Center
+      gridIndex: 237,
+      position: Vector(580, 420), // Center
     },
     {
       type: CornerType.BOTTOM_LEFT,
-      position: Vector(TWO_BY_TWO_CENTER_X, 700), // Center
+      gridIndex: 433,
+      position: Vector(580, 700), // Center
     },
     {
       type: CornerType.BOTTOM_RIGHT,
+      gridIndex: 447,
       position: Vector(1100, 700),
     },
   ],
@@ -257,27 +309,33 @@ export const ROOM_SHAPE_CORNERS: {
   [RoomShape.LBR]: [
     {
       type: CornerType.TOP_LEFT,
+      gridIndex: 0,
       position: Vector(60, 140),
     },
     {
       type: CornerType.TOP_RIGHT,
+      gridIndex: 27,
       position: Vector(1100, 140),
     },
     {
       type: CornerType.TOP_LEFT,
-      position: Vector(TWO_BY_TWO_CENTER_X, TWO_BY_TWO_CENTER_Y), // Center
+      gridIndex: 238,
+      position: Vector(580, 420), // Center
     },
     {
       type: CornerType.BOTTOM_RIGHT,
-      position: Vector(1100, TWO_BY_TWO_CENTER_Y), // Right
+      gridIndex: 251,
+      position: Vector(1100, 420), // Right
     },
     {
       type: CornerType.BOTTOM_LEFT,
+      gridIndex: 420,
       position: Vector(60, 700),
     },
     {
       type: CornerType.BOTTOM_RIGHT,
-      position: Vector(TWO_BY_TWO_CENTER_X, 700), // Right
+      gridIndex: 434,
+      position: Vector(580, 700), // Right
     },
   ],
 } as const;